move AboutUs related components to About folder
