+++ client/src/components/About/PeopleSay.jsx
